+++ src/main/java/objects/Enemy_system/EnemyAI/DefaultAI.java
@@ -1,33 +1,31 @@
-package Battle.Enemy;
+package objects.Enemy_system.EnemyAI;
 
 import java.util.Objects;
 import java.util.Random;
 
-public class SmartAI implements EnemyAI{
+public class DefaultAI implements EnemyAI{
+
     private EnemyInfo enemyInfo;
     private int attackChance;
 
-    public SmartAI(EnemyInfo enemyInfo, int attackChance){
+    public DefaultAI(EnemyInfo enemyInfo, int attackChance){
         this.enemyInfo = enemyInfo;
         this.attackChance = attackChance;
     }
 
     @Override
     public String respond(String input) {
-        if(this.enemyInfo.getHealth() < 10) {
-            return "use potion";
-        } else if (this.enemyInfo.getHealth() > 60) {
-            return "use skill";
-
-        } else {
+        if (Objects.equals(input, "use skill")) {
             Random rand = new Random();
             int upperbound = 101;
             int int_random = rand.nextInt(upperbound);
-            if (int_random < this.attackChance) {
+            if(int_random < this.attackChance){
                 return "use skill";
-            } else {
+            } else{
                 return "use potion";
             }
+        } else{
+            return "use skill";
         }
     }
 }
