+++ src/main/java/objects/Enemy_system/EnemyAI/EnemyAI.java
@@ -1,4 +1,4 @@
-package Battle.Enemy;
+package objects.Enemy_system.EnemyAI;
 
 public interface EnemyAI {
     public String respond(String input);
